+++ src/_cds/_cdsmodule.h
@@ -3,6 +3,10 @@
 
 #include <Python.h>
 
+#include "_cdscontext.h"
+#include "platforms.h"
+#include "pythoncapi_compat.h"
+
 // immortal & static objects
 #if PY_MINOR_VERSION >= 12
 #include <internal/pycore_long.h>
@@ -28,56 +32,16 @@
 #error Requires CPython 3.8+.
 #endif
 
-#include <stdbool.h>
-
 #if PY_MINOR_VERSION >= 12
 #include "clinic/_cdsmodule.c.h"
 #else
 #include "clinic/_cdsmodule-b4-312.c.h"
 #endif
-#include "lookup_table.h"
-#include "platforms.h"
-#include "pythoncapi_compat.h"
 
 #define CDS_MAX_IMG_SIZE (1024 * 1024 * 1024)
 #define CDS_REQUESTING_ADDR ((void *)0x280000000L)
 
 #define FAST_PATCH
-
-struct StringRefItem {
-    PyObject **ref;
-    struct StringRefItem *next;
-};
-
-struct StringRefList {
-    PyObject *str;
-    struct StringRefItem *refs;
-    struct StringRefList *next;
-};
-
-struct CDSArchiveHeader {
-    void *mapped_addr;
-    void *none_addr;
-    void *true_addr;
-    void *false_addr;
-    void *ellipsis_addr;
-    size_t used;
-
-    PyObject *obj;
-
-    struct StringRefList *all_string_ref;
-};
-
-struct MoveInContext {
-    int n_alloc;
-
-    table *orig_pyobject_to_in_heap_pyobject_map;
-    table *in_heap_str_to_string_ref_list_map;
-
-#if PY_MINOR_VERSION >= 12
-    PyObject *static_strings;
-#endif
-};
 
 /*
  * Internal representation of PYCDSMODE.
@@ -101,26 +65,6 @@ struct MoveInContext {
 #define P(p) ((p_type)(p))
 #define ALIEN_TO(size, align) (((size) + ((align)-1)) & ~((align)-1))
 #define UNSHIFT(p, shift, ty) ((ty *)(P(p) + (shift)))
-
-struct CDSStatus {
-    int verbose;
-    int mode;
-
-    // Prevent re-entry of cds.init_from_env()
-    bool initialized;
-
-    long shift;
-    bool traverse_error;
-
-    const char *archive;
-    fd_type archive_fd;
-
-    struct CDSArchiveHeader *archive_header;
-
-    PyObject *flags;
-
-    struct MoveInContext *move_in_ctx;
-};
 
 void *PyCDS_Malloc(size_t);
