+++ src/_cds/platforms.c
@@ -51,6 +51,38 @@ truncate_fd(fd_type fd, size_t size)
 #endif
 }
 
+#if IS_WINDOWS
+inline void
+debug_windows_err(LPCVOID lpAddress, DWORD size)
+{
+    if (GetLastError() != ERROR_INVALID_ADDRESS) {
+        verbose("unhandled error");
+    }
+
+    LPCVOID lpEndAddress = (LPCVOID)((DWORD_PTR)lpAddress + size);
+
+    MEMORY_BASIC_INFORMATION mbi;
+    bool canBeMapped = true;
+
+    while ((DWORD_PTR)lpAddress < (DWORD_PTR)lpEndAddress) {
+        if (VirtualQuery(lpAddress, &mbi, sizeof(mbi)) == 0) {
+            verbose("VirtualQuery failed with error: %p", GetLastError());
+            return;
+        }
+
+        if (mbi.State != MEM_FREE) {
+            canBeMapped = false;
+            verbose(
+                "Address range starting from %p of size %ld bytes is not "
+                "free. State: %d",
+                GetLastError(), mbi.RegionSize, mbi.State);
+        }
+
+        lpAddress = (LPCVOID)((DWORD_PTR)mbi.BaseAddress + mbi.RegionSize);
+    }
+}
+#endif
+
 void *
 create_map_from_archive(void *addr, size_t size, fd_type fd)
 {
@@ -64,10 +96,13 @@ create_map_from_archive(void *addr, size_t size, fd_type fd)
     HANDLE mapping =
         CreateFileMapping(fd, NULL, PAGE_READWRITE, 0, size, NULL);
     if (mapping == NULL) {
+        verbose("mapping from new archive failed: %p", GetLastError());
         goto fail;
     }
     res = MapViewOfFileEx(mapping, FILE_MAP_WRITE, 0, 0, 0, addr);
     if (res == NULL || res != addr) {
+        debug_windows_err(addr, size);
+        CloseHandle(mapping);
         goto fail;
     }
 #endif
@@ -109,38 +144,6 @@ fail:
     return NULL;
 }
 
-#if IS_WINDOWS
-inline void
-debug_windows_err(LPCVOID lpAddress, DWORD size)
-{
-    if (GetLastError() != ERROR_INVALID_ADDRESS) {
-        verbose("unhandled error");
-    }
-
-    LPCVOID lpEndAddress = (LPCVOID)((DWORD_PTR)lpAddress + size);
-
-    MEMORY_BASIC_INFORMATION mbi;
-    bool canBeMapped = true;
-
-    while ((DWORD_PTR)lpAddress < (DWORD_PTR)lpEndAddress) {
-        if (VirtualQuery(lpAddress, &mbi, sizeof(mbi)) == 0) {
-            verbose("VirtualQuery failed with error: %p", GetLastError());
-            return;
-        }
-
-        if (mbi.State != MEM_FREE) {
-            canBeMapped = false;
-            verbose(
-                "Address range starting from %p of size %ld bytes is not "
-                "free. State: %d",
-                GetLastError(), mbi.RegionSize, mbi.State);
-        }
-
-        lpAddress = (LPCVOID)((DWORD_PTR)mbi.BaseAddress + mbi.RegionSize);
-    }
-}
-#endif
-
 void *
 map_archive(fd_type file, size_t size, void *addr)
 {
@@ -155,6 +158,7 @@ map_archive(fd_type file, size_t size, void *addr)
     HANDLE mapping =
         CreateFileMapping(file, NULL, PAGE_READONLY, 0, size, NULL);
     if (mapping == NULL) {
+        verbose("load mapping from archive failed: %p", GetLastError());
         goto fail;
     }
     p_type *shm = MapViewOfFileEx(mapping, FILE_MAP_COPY, 0, 0, 0, addr);
@@ -172,13 +176,15 @@ fail:
 void
 close_archive(fd_type *file)
 {
+    if (file > 0) {
 #if IS_POSIX
-    close(*file);
-    *file = 0;
+        close(*file);
+        *file = 0;
 #elif IS_WINDOWS
-    CloseHandle(*file);
-    *file = 0;
+        CloseHandle(*file);
+        *file = 0;
 #endif
+    }
 }
 
 void
